refactor(handler): split step.go into step.go + responsestep.go
ackSignerStep and validateAckSignatureStep are ResponseStep types that
sit on the response pipeline, not the inbound request pipeline. Keeping
them in their own per-step files created an inconsistency with signStep
and validateSignStep, which live together in step.go.

Consolidate by introducing responsestep.go for all definition.ResponseStep
implementations (ackSignerStep, validateAckSignatureStep) and keeping
step.go for all definition.Step implementations. Test files mirror the
same split: responsestep_test.go and step_test.go. No logic changes.
